Move Crux specific classes to presto-crux
Moving crux analyzer specific classes to presto-crux module.
Also renaming native analyzer to crux analyzer.
